tools/brand-portrait/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portrait-brand-generator"
version = "0.1.0"
description = "MVP scaffold for a brand-consistent portrait generator."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "sqlalchemy>=2.0.37,<3.0.0",
  "psycopg[binary]>=3.2.3,<4.0.0",
  "redis>=5.2.1,<6.0.0",
  "pydantic>=2.10.4,<3.0.0",
  "pydantic-settings>=2.7.1,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "pillow>=11.1.0,<12.0.0",
  "boto3>=1.36.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
  "httpx>=0.28.1,<1.0.0",
  "ruff>=0.8.6,<1.0.0"
]
vision = [
  "opencv-python-headless>=4.10.0,<5.0.0",
  "insightface>=0.7.3,<1.0.0",
  "onnxruntime>=1.20.0,<2.0.0"
]
ml = [
  "torch>=2.5.1,<3.0.0",
  "diffusers>=0.32.0,<1.0.0",
  "transformers>=4.47.0,<5.0.0",
  "accelerate>=1.2.1,<2.0.0",
  "safetensors>=0.4.5,<1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
